Reactor: ThemeProvider
