+++ src/context/ThemeContext.ts
@@ -1,0 +1,6 @@
+import { ThemeContextType } from "../types"
+import { createContext } from "react";
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+    undefined,
+);
