+++ src/types/ThemeType.ts
@@ -1,1 +1,7 @@
 export type Theme = "light" | "dark";
+
+export interface ThemeContextType {
+    theme: Theme;
+    toggleTheme: () => void;
+}
+
